Changed 'int' to 'uint64_t' in decode metrics to prevent overflow

+++ src/decode.c
@@ -32,6 +32,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <regex.h>
 #include <htslib/khash.h>
 #include <cram/sam_header.h>
+#include <inttypes.h>
 
 #include "bamit.h"
 
@@ -94,7 +95,7 @@ typedef struct {
     char *lib;
     char *sample;
     char *desc;
-    int reads, pf_reads, perfect, pf_perfect, one_mismatch, pf_one_mismatch;
+    uint64_t reads, pf_reads, perfect, pf_perfect, one_mismatch, pf_one_mismatch;
 } bc_details_t;
 
 void free_bcd(void *entry)
@@ -281,19 +282,19 @@ static char *checkBarcodeQuality(char *bc_tag, bam1_t *rec, opts_t *opts)
     return newBarcode;
 }
 
-void writeMetricsLine(FILE *f, bc_details_t *bcd, opts_t *opts, int total_reads, int max_reads, int total_pf_reads, int max_pf_reads, int total_pf_reads_assigned, int nReads)
+void writeMetricsLine(FILE *f, bc_details_t *bcd, opts_t *opts, uint64_t total_reads, uint64_t max_reads, uint64_t total_pf_reads, uint64_t max_pf_reads, uint64_t total_pf_reads_assigned, uint64_t nReads)
 {
     fprintf(f, "%s\t", bcd->seq);
     fprintf(f, "%s\t", bcd->name);
     fprintf(f, "%s\t", bcd->lib);
     fprintf(f, "%s\t", bcd->sample);
     fprintf(f, "%s\t", bcd->desc);
-    fprintf(f, "%d\t", bcd->reads);
-    fprintf(f, "%d\t", bcd->pf_reads);
-    fprintf(f, "%d\t", bcd->perfect);
-    fprintf(f, "%d\t", bcd->pf_perfect);
-    fprintf(f, "%d\t", bcd->one_mismatch);
-    fprintf(f, "%d\t", bcd->pf_one_mismatch);
+    fprintf(f, "%"PRIu64"\t", bcd->reads);
+    fprintf(f, "%"PRIu64"\t", bcd->pf_reads);
+    fprintf(f, "%"PRIu64"\t", bcd->perfect);
+    fprintf(f, "%"PRIu64"\t", bcd->pf_perfect);
+    fprintf(f, "%"PRIu64"\t", bcd->one_mismatch);
+    fprintf(f, "%"PRIu64"\t", bcd->pf_one_mismatch);
     fprintf(f, "%f\t", total_reads ? bcd->reads / (double)total_reads : 0 );
     fprintf(f, "%f\t", max_reads ? bcd->reads / (double)max_reads : 0 );
     fprintf(f, "%f\t", total_pf_reads ? bcd->pf_reads / (double)total_pf_reads : 0 );
@@ -309,12 +310,12 @@ void writeMetricsLine(FILE *f, bc_details_t *bcd, opts_t *opts, int total_reads,
 int writeMetrics(va_t *barcodeArray, opts_t *opts)
 {
     bc_details_t *bcd = barcodeArray->entries[0];
-    int total_reads = bcd->reads;
-    int total_pf_reads = bcd->pf_reads;
-    int total_pf_reads_assigned = 0;
-    int max_reads = 0;
-    int max_pf_reads = 0;
-    int nReads = 0;
+    uint64_t total_reads = bcd->reads;
+    uint64_t total_pf_reads = bcd->pf_reads;
+    uint64_t total_pf_reads_assigned = 0;
+    uint64_t max_reads = 0;
+    uint64_t max_pf_reads = 0;
+    uint64_t nReads = 0;
     int n;
 
     // Open the metrics file
